pyproject.toml
[project]
name = "pliro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "google-genai>=1.74.0",
    "httpx>=0.28.1",
    "langchain==0.3.25",
    "langchain-core==0.3.65",
    "langchain-groq>=0.3.2",
    "langgraph==0.2.60",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]
